upstream(core): Preparations for a switchable storage backend (#10610)
# Description of change

- Upstream range: [v1.46.3, v1.47.1)
- Port commit: 
-
https://github.com/MystenLabs/sui/commit/e66ce415c6a160c0e6b577d84c496c72a94fcb41
- Description:

Preparation PR for a switchable storage backend.
 
Notes:
* `enum Storage` represents the storage backend with options for
RocksDB, InMemoryStore, and TH.
* the PR deprecates several unnecessary typed store methods and merges
redundant methods
* a basic InMemoryStore backend has been added as a reference
implementation, with potential future use in tests
* not all methods are covered. Iterators will be handled separately. The
returned interface will need to be changed, and additional methods will
be deprecated.
* the TH option is behind a feature flag. Extra options (e.g.,
with_rm_prefix) will be added in a followup
* the PR is essentially a noop and does not change the existing node's
behavior

### Additional fixes (review feedback)

* Generic types (`ColumnFamily`, `Storage`, `Database`, `GetResult`,
`StorageWriteBatch`) moved out of `rocks/mod.rs` into the new
`crates/typed-store/src/database.rs` module
* All visibility reduced to `pub(crate)` where external exposure is not
needed (`RocksDB`, `ColumnFamily`, `Storage`)


## Links to any relevant issues

Part of #10609

## How the change has been tested

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [ ] Patch-specific tests (correctness, functionality coverage)
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes
