+++ crates/typed-store/src/rocks/safe_iter.rs
@@ -8,14 +8,17 @@ use bincode::Options;
 use prometheus::{Histogram, HistogramTimer};
 use rocksdb::Direction;
 use serde::de::DeserializeOwned;
+use typed_store_error::TypedStoreError;
 
-use super::{RocksDBRawIter, TypedStoreError};
-use crate::metrics::{DBMetrics, RocksDBPerfContext};
+use crate::{
+    database::RawIter,
+    metrics::{DBMetrics, RocksDBPerfContext},
+};
 
 /// An iterator over all key-value pairs in a data map.
 pub struct SafeIter<'a, K, V> {
     cf_name: String,
-    db_iter: RocksDBRawIter<'a>,
+    db_iter: RawIter<'a>,
     _phantom: PhantomData<(K, V)>,
     direction: Direction,
     is_initialized: bool,
@@ -29,9 +32,9 @@ pub struct SafeIter<'a, K, V> {
 }
 
 impl<'a, K: DeserializeOwned, V: DeserializeOwned> SafeIter<'a, K, V> {
-    pub(super) fn new(
+    pub(crate) fn new(
         cf_name: String,
-        db_iter: RocksDBRawIter<'a>,
+        db_iter: RawIter<'a>,
         _timer: Option<HistogramTimer>,
         _perf_ctx: Option<RocksDBPerfContext>,
         bytes_scanned: Option<Histogram>,
