+++ crates/typed-store/src/rocks/tests.rs
@@ -2,12 +2,16 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+use std::ops::RangeBounds;
+
 use rstest::rstest;
+use serde::{Serialize, de::DeserializeOwned};
 
 use super::*;
 use crate::{
     reopen,
     rocks::safe_iter::{SafeIter, SafeRevIter},
+    traits::Map,
 };
 
 fn temp_dir() -> std::path::PathBuf {
@@ -90,7 +94,7 @@ async fn test_reopen() {
             .expect("Failed to insert");
         db
     };
-    let db = DBMap::<u32, String>::reopen(&arc.rocksdb, None, &ReadWriteOptions::default(), false)
+    let db = DBMap::<u32, String>::reopen(&arc.db, None, &ReadWriteOptions::default(), false)
         .expect("Failed to re-open storage");
     assert!(
         db.contains_key(&123456789)
@@ -116,18 +120,11 @@ async fn test_reopen_macro() {
     let keys_vals_cf1 = (1..100).map(|i| (i, i.to_string()));
     let keys_vals_cf2 = (1..100).map(|i| (i, i.to_string()));
 
-    assert_eq!(db_map_1.cf, FIRST_CF);
-    assert_eq!(db_map_2.cf, SECOND_CF);
+    assert_eq!(db_map_1.cf_name(), FIRST_CF);
+    assert_eq!(db_map_2.cf_name(), SECOND_CF);
 
     assert!(db_map_1.multi_insert(keys_vals_cf1).is_ok());
     assert!(db_map_2.multi_insert(keys_vals_cf2).is_ok());
-}
-
-#[tokio::test]
-async fn test_wrong_reopen() {
-    let rocks = open_rocksdb(temp_dir(), &["foo", "bar", "baz"]);
-    let db = DBMap::<u8, u8>::reopen(&rocks, Some("quux"), &ReadWriteOptions::default(), false);
-    assert!(db.is_err());
 }
 
 #[tokio::test]
@@ -395,14 +392,7 @@ async fn test_insert_batch_across_different_db() {
 
 #[tokio::test]
 async fn test_delete_batch() {
-    let db = DBMap::<i32, String>::open(
-        temp_dir(),
-        MetricConf::default(),
-        None,
-        None,
-        &ReadWriteOptions::default(),
-    )
-    .expect("Failed to open storage");
+    let db = open_map::<_, u32, String>(temp_dir(), None);
 
     let keys_vals = (1..100).map(|i| (i, i.to_string()));
     let mut batch = db.batch();
@@ -425,14 +415,14 @@ async fn test_delete_batch() {
 
 #[tokio::test]
 async fn test_delete_range() {
-    let db: DBMap<i32, String> = DBMap::open(
-        temp_dir(),
-        MetricConf::default(),
+    let options = ReadWriteOptions::default().set_ignore_range_deletions(false);
+    let db: DBMap<i32, String> = DBMap::reopen(
+        &open_rocksdb(temp_dir(), &[rocksdb::DEFAULT_COLUMN_FAMILY_NAME]),
         None,
-        None,
-        &ReadWriteOptions::default().set_ignore_range_deletions(false),
+        &options,
+        false,
     )
-    .expect("Failed to open storage");
+    .unwrap();
 
     // Note that the last element is (100, "100".to_owned()) here
     let keys_vals = (0..101).map(|i| (i, i.to_string()));
@@ -460,14 +450,7 @@ async fn test_delete_range() {
 
 #[tokio::test]
 async fn test_clear() {
-    let db = DBMap::<i32, String>::open(
-        temp_dir(),
-        MetricConf::default(),
-        None,
-        Some("table"),
-        &ReadWriteOptions::default(),
-    )
-    .expect("Failed to open storage");
+    let db: DBMap<i32, String> = open_map(temp_dir(), Some("table"));
     // Test clear of empty map
     let _ = db.unsafe_clear();
 
@@ -599,14 +582,7 @@ async fn test_range_iter() {
 
 #[tokio::test]
 async fn test_is_empty() {
-    let db = DBMap::<i32, String>::open(
-        temp_dir(),
-        MetricConf::default(),
-        None,
-        Some("table"),
-        &ReadWriteOptions::default(),
-    )
-    .expect("Failed to open storage");
+    let db: DBMap<i32, String> = open_map(temp_dir(), Some("table"));
 
     // Test empty map is truly empty
     assert!(db.is_empty());
@@ -658,7 +634,7 @@ async fn test_checkpoint() {
     db.multi_insert(keys_vals.clone())
         .expect("Failed to multi-insert");
     let checkpointed_path = path_prefix.join("checkpointed_db");
-    db.rocksdb
+    db.db
         .checkpoint(&checkpointed_path)
         .expect("Failed to create db checkpoint");
     // Create more kv pairs
@@ -719,14 +695,7 @@ async fn open_as_secondary_test() {
     let primary_path = temp_dir();
 
     // Init a DB
-    let primary_db = DBMap::<i32, String>::open(
-        primary_path.clone(),
-        MetricConf::default(),
-        None,
-        Some("table"),
-        &ReadWriteOptions::default(),
-    )
-    .expect("Failed to open storage");
+    let primary_db: DBMap<i32, String> = open_map(primary_path.clone(), Some("table"));
     // Create kv pairs
     let keys_vals = (0..101).map(|i| (i, i.to_string()));
 
@@ -736,11 +705,11 @@ async fn open_as_secondary_test() {
 
     let opt = rocksdb::Options::default();
     let secondary_store = open_cf_opts_secondary(
-        primary_path,
+        primary_path.clone(),
         None,
         None,
         MetricConf::default(),
-        &[("table", opt)],
+        &[("table", opt.clone())],
     )
     .unwrap();
     let secondary_db = DBMap::<i32, String>::reopen(
@@ -771,16 +740,16 @@ async fn open_as_secondary_test() {
 }
 
 fn open_map<P: AsRef<Path>, K, V>(path: P, opt_cf: Option<&str>) -> DBMap<K, V> {
-    DBMap::<K, V>::open(
-        path,
-        MetricConf::default(),
-        None,
+    let cf_key = opt_cf.unwrap_or(rocksdb::DEFAULT_COLUMN_FAMILY_NAME);
+    DBMap::<K, V>::reopen(
+        &open_rocksdb(path, &[cf_key]),
         opt_cf,
         &ReadWriteOptions::default(),
+        false,
     )
     .expect("failed to open rocksdb")
 }
 
-fn open_rocksdb<P: AsRef<Path>>(path: P, opt_cfs: &[&str]) -> Arc<RocksDB> {
+fn open_rocksdb<P: AsRef<Path>>(path: P, opt_cfs: &[&str]) -> Arc<Database> {
     open_cf(path, None, MetricConf::default(), opt_cfs).expect("failed to open rocksdb")
 }
